depend on wireguard-tools instead of wireguard meta package
Now that the wireguard module is integrated into the kernel used by Debian (since version 5.5) , we only need to install wireguard-tools. Moreover installing the wireguard meta package on some ARM boards running armbian pulls as a dependency an incompatible kernel and that results in a system that no longer boots.